lcfiplus processor should work with multiple instances single track pseudo-vertex added jet vertex refiner algorithm added bugfix in flavor tagging new flavor tagging variables

+++ LCFIPlus/trunk/include/algoSigProb.h
@@ -9,17 +9,17 @@
 namespace lcfiplus{
 namespace algoSigProb{
 
-extern float trackD0Significance(const Track* trk, const Vertex* pri);
-extern float trackZ0Significance(const Track* trk, const Vertex* pri);
-extern float signedD0Significance(const Track* trk, const Jet* jet, const Vertex* pri, bool updateFlt=false);
-extern float signedZ0Significance(const Track* trk, const Jet* jet, const Vertex* pri, bool updateFlt=false);
-extern void findMostSignificantTrack(const Jet* jet, const Vertex* pri, float sigVec[6]);
+extern double trackD0Significance(const Track* trk, const Vertex* pri);
+extern double trackZ0Significance(const Track* trk, const Vertex* pri);
+extern double signedD0Significance(const Track* trk, const Jet* jet, const Vertex* pri, bool updateFlt=false);
+extern double signedZ0Significance(const Track* trk, const Jet* jet, const Vertex* pri, bool updateFlt=false);
+extern void findMostSignificantTrack(const Jet* jet, const Vertex* pri, double sigVec[6]);
 
-extern float prob1D(float sig, float maxsig, float* pars);
-extern float trackProbD0(const Track* trk, const Vertex* pri);
-extern float trackProbZ0(const Track* trk, const Vertex* pri);
-extern float jointProbD0(const Jet* jet, const Vertex* pri);
-extern float jointProbZ0(const Jet* jet, const Vertex* pri);
+extern double prob1D(double sig, double maxsig, double* pars);
+extern double trackProbD0(const Track* trk, const Vertex* pri);
+extern double trackProbZ0(const Track* trk, const Vertex* pri);
+extern double jointProbD0(const Jet* jet, const Vertex* pri);
+extern double jointProbZ0(const Jet* jet, const Vertex* pri);
 
 }}
